pyproject.toml
[project]
name = "tiny-dreamer-highway"
version = "0.1.0"
description = "Test-first Tiny Dreamer scaffold for CSC 580 highway-env world models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gymnasium>=1.0.0",
    "highway-env>=1.9.0",
    "imageio>=2.37.0",
    "matplotlib>=3.10.0",
    "numpy>=2.2.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "torch>=2.6.0",
    "typing-extensions>=4.12.0",
]

[project.scripts]
tdh-collect-random = "tiny_dreamer_highway.cli:main"

[dependency-groups]
dev = [
    "ipykernel>=6.29.0",
    "pytest>=8.3.0",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.hatch.build.targets.wheel]
packages = ["src/tiny_dreamer_highway"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
